pyproject.toml
[project]
name = "ai-tutor-backend"
version = "0.1.0"
description = "Multi-tenant AI Tutor API"
authors = [
    {name = "Tanishk",email = "tanishk.soni@adda247.com"}
]
readme = "README.md"
requires-python = "3.10.0"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "qdrant-client (>=1.16.2,<2.0.0)",
    "pymupdf (>=1.26.7,<2.0.0)",
    "google-generativeai (>=0.8.0,<1.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<1.0.0)",
    "email-validator (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
